requirements.txt
aiogram>=3.4.0
APScheduler==3.10.4
pytz==2024.1
python-dotenv==1.0.1
aiofiles>=23.2.1